power: smb135x: Make the code 64 bit compatible
As the pointers size change to 64 bits in the 64 bit kernel, the
int declarations for them from the legacy code give compilation
warnings.

Fix this by replace int casting of pointers with long

Reviewed-on: http://gerrit.mot.com/702753
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

+++ drivers/power/smb135x-charger-mmi.c
@@ -4437,7 +4437,7 @@ static int smb_parse_dt(struct smb135x_chg *chip)
 		return -EINVAL;
 	}
 
-	chip->usb_current_arr_size = (int)match->data;
+	chip->usb_current_arr_size = (long)match->data;
 
 	rc = of_property_read_u32(node, "qcom,float-voltage-mv",
 						&chip->vfloat_mv);
